deletion method updated

+++ src/main/java/com/epam/lowcost/service/implementations/FlightServiceImpl.java
@@ -43,9 +43,10 @@ public class FlightServiceImpl implements FlightService {
 
     @Override
     public Flight deleteFlight(Long id) {
-        ticketDAO.deleteTicketsByFlightId(id);
-        return flightDAO.deleteFlight(id);
-
+        if (ticketDAO.deleteTicketsByFlightId(id)) {
+            return flightDAO.deleteFlight(id);
+        }
+        return null;
     }
 
     @Override
