+++ src/main/java/com/epam/lowcost/service/implementations/UserServiceImpl.java
@@ -6,7 +6,6 @@ import com.epam.lowcost.model.User;
 import com.epam.lowcost.service.interfaces.UserService;
 import org.mindrot.jbcrypt.BCrypt;
 
-import java.awt.*;
 import java.util.List;
 
 public class UserServiceImpl implements UserService {
@@ -43,8 +42,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public String deleteUser(long userId) {
-        ticketDAO.deleteTicketsByUserId(userId);
-        return userDAO.deleteUser(userId);
+        if (ticketDAO.deleteTicketsByUserId(userId)) {
+            return userDAO.deleteUser(userId);
+        }
+        return null;
     }
 
     @Override
